+++ packages/cli/src/commands/network/contracts.ts
@@ -10,6 +10,9 @@ const UNVERSIONED_CONTRACTS = [
   CeloContract.Registry,
   CeloContract.FeeCurrencyWhitelist,
   CeloContract.Freezer,
+  CeloContract.StableToken,
+  CeloContract.StableTokenBRL,
+  CeloContract.StableTokenEUR,
 ]
 const UNPROXIED_CONTRACTS: CeloContract[] = []