+++ packages/cli/src/utils/off-chain-data.ts
@@ -46,10 +46,7 @@ export abstract class OffchainDataCommand extends BaseCommand {
     }: ParserOutput<any, any> = this.parse()
 
     const from = privateKeyToAddress(privateKey)
-    // identity depends on a version of kit that doesnt support cip64 tx
-    // so the types dont match but it is fine to ignore here as the kit is not sending a tx
-    // its reading from AccountsWrapper
-    // @ts-expect-error
+    // @ts-ignore -- TODO: if identity depends on diff version of ck which has a slightly differnt type this complains
     this.offchainDataWrapper = new BasicDataWrapper(from, this.kit)
 
     this.offchainDataWrapper.storageWriter =