+++ packages/protocol/truffle-config.js
@@ -2,11 +2,15 @@ const SOLC_VERSION = '0.5.13'
 
 const parent = require('./truffle-config-parent.js')
 const networks = { ...parent.networks }
+const { celoScanApiKey } = require('./.env.json')
 
 console.log(`Using truffle version for Solidity ${SOLC_VERSION}`)
 
 module.exports = {
-  plugins: ['truffle-security', 'truffle-plugin-blockscout-verify'],
+  plugins: ['truffle-plugin-verify'],
+  api_keys: {
+    celoscan: celoScanApiKey,
+  },
   compilers: {
     solc: {
       version: SOLC_VERSION,
@@ -30,7 +34,10 @@ if (process.argv.includes('--gas')) {
         },
       },
     },
-    plugins: ['truffle-security', 'truffle-plugin-blockscout-verify'],
+    plugins: ['truffle-plugin-verify'],
+    api_keys: {
+      celoscan: celoScanApiKey,
+    },
     networks,
     reporter: 'eth-gas-reporter',
     reporterOptions: {