+++ packages/sdk/contractkit/src/wrappers/MetaTransactionWalletDeployer.test.ts
@@ -1,16 +1,16 @@
 import { Address, normalizeAddress } from '@celo/base'
+import MTWContractArtifacts from '@celo/celo-devchain/contracts/contracts-0.5/MetaTransactionWallet.json'
+import MTWDeployerContractArtifacts from '@celo/celo-devchain/contracts/contracts-0.5/MetaTransactionWalletDeployer.json'
 import { CeloTxReceipt, EventLog } from '@celo/connect'
 import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
-import MTWContract from '@celo/protocol/build/contracts/MetaTransactionWallet.json'
-import MTWDeployerContract from '@celo/protocol/build/contracts/MetaTransactionWalletDeployer.json'
 import { MetaTransactionWallet, newMetaTransactionWallet } from '../generated/MetaTransactionWallet'
 import { newProxy } from '../generated/Proxy'
 import { newKitFromWeb3 } from '../kit'
 import { MetaTransactionWalletDeployerWrapper } from './MetaTransactionWalletDeployer'
 
 const contract = require('@truffle/contract')
-const MetaTransactionWalletDeployer = contract(MTWDeployerContract)
-const MetaTransactionWallet = contract(MTWContract)
+const MetaTransactionWalletDeployer = contract(MTWDeployerContractArtifacts)
+const MetaTransactionWallet = contract(MTWContractArtifacts)
 
 testWithGanache('MetaTransactionWallet Wrapper', (web3) => {
   MetaTransactionWalletDeployer.setProvider(web3.currentProvider)