+++ packages/sdk/contractkit/src/wrappers/MetaTransactionWalletDeployer.ts
@@ -1,7 +1,13 @@
 import { MetaTransactionWalletDeployer } from '../generated/MetaTransactionWalletDeployer'
 import { BaseWrapper, proxySend } from './BaseWrapper'
 
+/*
+ * @deprecated https://github.com/celo-org/celo-monorepo/issues/10766
+ */
 export class MetaTransactionWalletDeployerWrapper extends BaseWrapper<MetaTransactionWalletDeployer> {
+  /*
+   * @deprecated https://github.com/celo-org/celo-monorepo/issues/10766
+   */
   deploy = proxySend(this.connection, this.contract.methods.deploy)
 }