+++ scripts/hooks/pre-push.js
@@ -60,7 +60,13 @@ function getDateFromFirstCommit(fromSHA, toSHA) {
 ////////////////////////////////////////////////////////////////
 
 // must trim otherwise the name will be 'origin\n'
+
 const remoteName = execSync('git remote').toString().trim()
+// create remote tracking branches to ensure we can compare current to origin/master
+const remoteTrackingMasterBranches = execSync('git ls-remote --heads origin master')
+if (remoteTrackingMasterBranches.toString().trim() === '') {
+  execSync('git fetch origin')
+}
 const changes = process.env.HUSKY_GIT_STDIN.split('\n')
   .filter((line) => line !== '')
   .map((line) => {