+++ packages/sdk/governance/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @celo/governance
 
+## 5.0.6
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [679ef0c60]
+- Updated dependencies [32face3d8]
+- Updated dependencies [97d5ccf43]
+- Updated dependencies [87647b46b]
+  - @celo/contractkit@5.2.0
+  - @celo/connect@5.1.1
+  - @celo/base@6.0.0
+  - @celo/explorer@5.0.6
+  - @celo/utils@5.0.6
+
+## 5.0.6-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [32face3d8]
+- Updated dependencies [97d5ccf43]
+- Updated dependencies [87647b46b]
+  - @celo/contractkit@5.2.0-beta.0
+  - @celo/base@6.0.0-beta.0
+  - @celo/explorer@5.0.6-beta.0
+  - @celo/connect@5.1.1-beta.0
+  - @celo/utils@5.0.6-beta.0
+
 ## 5.0.5
 
 ### Patch Changes