+++ packages/sdk/transactions-uri/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @celo/transactions-uri
 
+## 5.0.6
+
+### Patch Changes
+
+- Updated dependencies [679ef0c60]
+- Updated dependencies [97d5ccf43]
+  - @celo/connect@5.1.1
+  - @celo/base@6.0.0
+
+## 5.0.6-beta.0
+
+### Patch Changes
+
+- Updated dependencies [97d5ccf43]
+  - @celo/base@6.0.0-beta.0
+  - @celo/connect@5.1.1-beta.0
+
 ## 5.0.5
 
 ### Patch Changes